JDG-1087 ISPN-7296 Add eviction for off-heap
* Fixed lru count not including hashCode
* Moved StorageType to public package
* Updated schema for some additional documentation
* Reduced load for test so it doesn't fail with tracing

+++ core/src/main/java/org/infinispan/container/offheap/UnpooledOffHeapMemoryAllocator.java
@@ -31,8 +31,8 @@ public class UnpooledOffHeapMemoryAllocator implements OffHeapMemoryAllocator {
       long memoryLocation = UNSAFE.allocateMemory(memoryLength);
       long currentSize = amountAllocated.addAndGet(memoryLength);
       if (trace) {
-         log.tracef("Allocated off heap memory at %d with %d bytes.  Total size: %d", memoryLocation, memoryLength,
-               currentSize);
+         log.tracef("Allocated off heap memory at %d with %d bytes.  Total size: %d", (Long) memoryLocation, (Long) memoryLength,
+               (Long) currentSize);
       }
       return memoryLocation;
    }
@@ -46,8 +46,8 @@ public class UnpooledOffHeapMemoryAllocator implements OffHeapMemoryAllocator {
    public void deallocate(long memoryAddress, long size) {
       long currentSize = amountAllocated.addAndGet(- size);
       if (trace) {
-         log.tracef("Deallocating off heap memory at %d with %d bytes.  Total size: %d", memoryAddress, size,
-               currentSize);
+         log.tracef("Deallocating off heap memory at %d with %d bytes.  Total size: %d", (Long) memoryAddress, (Long) size,
+               (Long) currentSize);
       }
       UNSAFE.freeMemory(memoryAddress);
    }
